Don't handle DESTROY with AUTOLOAD
When the last reference to a Supervisord::Client object goes away, its
`DESTROY` method is called.  Because no `DESTROY` method is defined, its
`AUTOLOAD` method handles it instead.  This results in an RPC call to
`supervisor.DESTROY` - a method which does not exist in Supervisor.
This could cause problems if such a method is ever implemented.  It also
causes problems if, for example, the RPC calls are made through an SSH
tunnel which is destroyed at object destruction, before the
Supervisord::Client object is destroyed.